requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
cachetools==4.2.2
captum==0.4.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
configparser==5.0.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.3
fsspec==2021.5.0
future==0.18.2
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.0
google-auth-oauthlib==0.4.4
greenlet==1.1.0
grpcio==1.38.0
idna==2.10
imageio==2.9.0
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
itk==5.1.2
itk-core==5.1.2
itk-filtering==5.1.2
itk-io==5.1.2
itk-numerics==5.1.2
itk-registration==5.1.2
itk-segmentation==5.1.2
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
monai==0.5.2
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nibabel==3.2.1
notebook==6.4.0
numba==0.53.1
numpy==1.20.3
oauthlib==3.1.0
opencv-python==4.5.3.56
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pluggy==0.13.1
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.0
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyDeprecate==0.3.0
pydicom==2.1.2
Pygments==2.9.0
pylidc==0.2.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
pytorch-lightning==1.3.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
sentry-sdk==1.1.0
shap==0.39.0
shortuuid==1.0.1
six==1.16.0
slicer==0.0.7
smmap==4.0.0
SQLAlchemy==1.4.15
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
torch==1.8.1
torchmetrics==0.3.2
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.4
wandb==0.10.30
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
yarl==1.6.3
